fix: making use popper universal (#371)
* staging work

* refactor(popper): making usePopper portable
