+++ website/play/index.vue
@@ -1,6 +1,17 @@
 <template>
-  <div>
-  </div>
+  <el-popper
+    placement="bottom-end"
+    effect="light"
+    pure
+    trigger="click"
+  >
+    <template #trigger>
+      <el-button>Button</el-button>
+    </template>
+    <div>
+      I am the content
+    </div>
+  </el-popper>
 </template>
 
 <script>
